requirements.txt
python-telegram-bot>=21.0.1
gspread>=5.12.0
oauth2client>=4.1.3
pydantic>=2.7
python-dotenv>=1.0.1
aiogram>=3.20.0.post0,<4.0.0
google-auth>=2.40.3,<3.0.0
google-auth-oauthlib>=1.2.2,<2.0.0
google-api-python-client>=2.173.0,<3.0.0
